move gumbel loss into utils
